chapter7/pyproject.toml
[tool.poetry]
name = "agent-book"
version = "0.1.0"
description = ""
authors = ["Your Name <user@example.com>"]
readme = "README.md"
packages = [{include = "agent_book"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.53.1"
langgraph = "^0.2.14"
pydantic-settings = "^2.4.0"
langchain = "^0.3.0" 
promptflow-tracing = "^1.16.1"
promptflow-devkit = "^1.16.1"
langchain-openai = "^0.3.6"
jira = "^3.8.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "F", # Flake8
    "B", # Black
    "I", # isort
    "E", # error
    "W"  # warning
]
ignore = []
fixable = ["ALL"]
exclude = [
    ".ruff_cache"
]
line-length = 119
indent-width = 4
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
force-single-line = false
